fix: port fixes from llada2-support branch
- Attention mask: use 1/0 float instead of 0/-inf (matches official LLaDA2 code)
- Sampling: add shift-right in top-p to preserve at least one token
- Sampling: apply temperature before top-k/top-p filtering
- Sampling: separate greedy branch for temperature=0
- LLaDA2Pipeline: update defaults to LLaDA2.1 quality mode (threshold=0.7,
  editing_threshold=0.5, max_post_steps=16, eos_early_stop=True)
- LLaDA2Pipeline: remove duplicate _prepare_prompt_ids, reuse mixin's
  _prepare_input_ids
- LLaDA2Pipeline: fix EXAMPLE_DOC_STRING (model ID, dtype, scheduler)
- Simplify _extract_input_ids since return_dict=True
- Replace deprecated torch_dtype with dtype in example scripts
- Add all discrete diffusion pipeline/scheduler pages to docs toctree

+++ examples/discrete_diffusion/sample_llada2.py
@@ -172,7 +172,7 @@ def main():
         model = AutoModelForCausalLM.from_pretrained(
             args.model_id,
             trust_remote_code=True,
-            torch_dtype=torch_dtype,
+            dtype=torch_dtype,
             low_cpu_mem_usage=True,
         )
         # Apply group offloading with CUDA streams for better performance
@@ -191,7 +191,7 @@ def main():
         model = AutoModelForCausalLM.from_pretrained(
             args.model_id,
             trust_remote_code=True,
-            torch_dtype=torch_dtype,
+            dtype=torch_dtype,
             low_cpu_mem_usage=True,
         )
         # Sequential offloading will be applied via pipeline
@@ -200,7 +200,7 @@ def main():
         model = AutoModelForCausalLM.from_pretrained(
             args.model_id,
             trust_remote_code=True,
-            torch_dtype=torch_dtype,
+            dtype=torch_dtype,
             device_map="auto",
             low_cpu_mem_usage=True,
         )
